BREAKING CHANGES!
Fixed #53: Extra fields removed
Modified code to be full based on the debug Panels... So far 2 have been implemented: Request and Db, the rest will follow soon.
To also be able to display internal data there is a "view only panel" concept which still needs to be improved
Callback functionality changed for the user. It's now called userIdentifier, since not everyone is going to use a username. Could be an email etc.