cURL support and easier panel configuration
Implemented the cURL panel to allow you to track your cURL requests as described in #65. Keeping the issue open as we still need to add tests.
Modified the panel configuration. There now is an extra `$panelsMerge` option that is merged together with the regular configuration, for quicker inclusion of a new panel or just modifying configuration somewhat. This closes #96 but needs more tests.